requirements.txt
@"
oracledb
python-dotenv
requests
pandas
"@ | Set-Content -Encoding utf8 requirements.txt
